+++ frontend/src/components/TicketDetailsRightDrawer.tsx
@@ -19,6 +19,7 @@ import EpicLinkEditItem from "./TicketRightDrawerItems/EpicLinkEditItem";
 import StoryPointsEditItem from "./TicketRightDrawerItems/StoryPointsEditItem";
 import DeleteTicketDialog from "./DeleteTicketDialog";
 import { TICKET_DRAWER_WIDTH } from "../utils/constants";
+import { Toolbar } from "@mui/material";
 
 interface TicketDetailsRightDrawerProps {
   ticket: TicketInterface;
@@ -40,8 +41,17 @@ const TicketDetailsRightDrawer = (props: TicketDetailsRightDrawerProps): JSX.Ele
       anchor='right'
       open={isOpen}
       variant='persistent'
-      sx={{ width: TICKET_DRAWER_WIDTH, flexShrink: 0, "& .MuiDrawer-paper": { width: TICKET_DRAWER_WIDTH } }}
+      sx={{
+        width: TICKET_DRAWER_WIDTH,
+        flexShrink: 0,
+        [`& .MuiDrawer-paper`]: {
+          width: TICKET_DRAWER_WIDTH,
+          boxSizing: "border-box",
+          zIndex: 0,
+        },
+      }}
     >
+      <Toolbar />
       <List>
         <ListItem disablePadding sx={{ pl: 1, py: 1 }}>
           <IconButton size='small' onClick={onClose}>