+++ frontend/src/components/TicketRightDrawerItems/AssigneeEditItem.tsx
@@ -84,9 +84,8 @@ const AssigneeEditItem = (props: AssigneeEditItemProps): JSX.Element => {
   if (isEditModeOn) {
     return (
       <ListItem sx={{ display: "flex", flexDirection: "column", alignItems: "flex-start" }}>
-        <Box display='flex' width='100%' mb={1.5} gap={2}>
+        <Box display='flex' width='100%' mb={2} gap={1}>
           <Typography variant='body2'>Assignee</Typography>
-          <Box flexGrow={1} />
           <UpdateButton
             onAccept={handleUpdateTicketAssignee}
             onCancel={handleToggleEditMode}
@@ -156,7 +155,7 @@ const AssigneeEditItem = (props: AssigneeEditItemProps): JSX.Element => {
 
   return (
     <ListItem sx={{ display: "flex", flexDirection: "column", alignItems: "flex-start" }}>
-      <Box display='flex' justifyContent='space-between' width='100%' mb={1}>
+      <Box display='flex' width='100%' gap={1} mb={2}>
         <Typography variant='body2'>Assignee</Typography>
         <EditButton onStartEdit={handleToggleEditMode} />
       </Box>