+++ frontend/src/components/TicketRightDrawerItems/DescriptionEditItem.tsx
@@ -39,7 +39,7 @@ const DescriptionEditItem = (props: DescriptionEditItemProps): JSX.Element => {
   if (!isEditModeOn) {
     return (
       <ListItem sx={{ display: "flex", flexDirection: "column", alignItems: "flex-start" }}>
-        <Box display='flex' justifyContent='space-between' width='100%' mb={0.5}>
+        <Box display='flex' gap={1} width='100%' mb={0.5}>
           <Typography variant='body2'>Notes</Typography>
           <EditButton onStartEdit={handleToggleEditMode} />
         </Box>
@@ -51,9 +51,8 @@ const DescriptionEditItem = (props: DescriptionEditItemProps): JSX.Element => {
 
   return (
     <ListItem sx={{ display: "flex", flexDirection: "column", alignItems: "flex-start" }}>
-      <Box display='flex' width='100%' mb={0.5} gap={2}>
+      <Box display='flex' width='100%' mb={0.5} gap={1}>
         <Typography variant='body2'>Notes</Typography>
-        <Box flexGrow={1} />
         <UpdateButton onAccept={handleUpdateTicket} onCancel={handleToggleEditMode} showUpdateButton />
       </Box>
       <TextField